[PATCH] gianfar mii needs to zero out the mii_bus structure
To ensure that phy_mask and any future elements of the mii_bus
structure are initialized use kzalloc() instead of kmalloc().

This fixes an issue in which phy_mask was not being initialized
and we would skip random phy addresses when scanning.

Signed-off-by: Kumar Gala <galak@kernel.crashing.org>

From: Kumar Gala <galak@gate.crashing.org>
Date: Wed, 9 Nov 2005 12:13:11 -0600 (CST) (19:13 MET)
To: jgarzik@pobox.com
Cc: netdev@vger.kernel.org, linuxppc-embedded@ozlabs.org