[POWERPC] Make FEC work on the lite5200
This patch may very well break Eth support on the Efika, and it's not
very pretty.  But is works well enough for an NFS rootfs.  This also
makes major bestcomm changes by removing Efika-specific bestcomm
support and porting the arch/ppc bestcomm support driver which
was posted to linuxppc-embedded about a year ago.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>