[PATCH] ibm_emac: Add support for interrupt coalescing (440EPx/440GRx)
This patch adds support for EMAC interrupt coalescing for the AMCC 440EPx
and 440GRx PPC's to the ibm_emac ethernet driver used by all 4xx PPC's.
For certain applications, ethernet interrupt coalescing may increase the
overall throughput.

The parameters (rx-/tx-frames and timeout) can be configured using the
Linux build system and changed in the running system using ethtool.

For example by using the following command:

$ ethtool -C eth0 rx-frames 8 rx-usecs 5000 tx-frames 8 tx-usecs 5000

One restriction right now is, that when enabled in the Linux kernel, all
EMAC interfaces will run in the interrupt coalescing mode. This should not
be a problem since each interface can be configured individually. So eth1
could be configured with 1 rx-/tx-frames and without timeout (same as non
interrupt coalescing mode), and eth0 could be configured as described in
the example above.

Signed-off-by: Stefan Roese <sr@denx.de>