ppc: Add support for bigger page sizes than 4KB on PPC44x
This patch adds support for the PAGE_SIZE set equal to 64 KB on
PPC32 platforms. Increasing the PAGE_SIZE may improve the performance
of the system in some cases.

RAID processing with dedicated hardware that offloads the CPU is one
of such cases. Using the bigger PAGE_SIZE can decrease the number of
CPU operations required to process the data blocks, decrease the
control overhead and so improve the overall performance. For example,
changing the PAGE_SIZE from 4KB to 64KB on a Yucca (440SPe) gives
~300% performance benefit for full-stripe writing in a (7d+1p) RAID-5
setup.

Signed-off-by: Yuri Tikhonov <yur@emcraft.com>
Acked-by: Wolfgang Denk <wd@denx.de>
Acked-by: Stefan Roese <sr@denx.de>