RAID: avoid memory copy to optimize performance
The patch adds an option to skip intermediate data copying between
the bio requested to write and the disk cache in <sh> if the
full-stripe write operation is in progress. This might improve the
performance of write operations in some dedicated cases but generally
eliminating disk cache slows the performance down.

Signed-off-by: Yuri Tikhonov <yur@emcraft.com>
--