USB: Fix OHCI driver for IBM/AMCC 4xx PPC's
This patch makes the OHCI code work correctly on AMCC PPC440
cpus with internal USB-host controller. It seems that the
optimization is not allowed for this OHCI implementation.

This patch is slightly modified by Stefan Roese, to not only
make this fix for 440EP & 440EPx, but for all 4xx PPC's, since
we already have another 4xx (405EZ) with this OHCI
implementation and others are in the queue.

Signed-off-by: Gerhard Jaeger <gjaeger@sysgo.com>
Signed-off-by: Stefan Roese <sr@denx.de>