[PATCH] mpc52xx: Correct calculation of FEC RX errors.
'ifconfig eth0' command for mpc5200B-based cards shows error for RX.
However none of RX MIB counters is set to value greater than zero.
Number of errors is equal to number of multicast packet. In linux 2.4
calculation of RX errors is slightly different and takes into account
number of multicast packet. This change is a port of calculation method
of RX errors for FEC controller from linux 2.4 to 2.6.

Signed-off-by: Grzegorz Bernacki <gjb@semihalf.com>