ppc: Fix problem with 44x L2-cache IRQ number used in ibm440gx_common.c
The routine ibm440gx_l2c_enable() is now used not only from 440GX but also
from 440SP/440SPe PPC's. But is still used a fixed IRQ number for the L2
cache interrupt which is different on 440SP/440SPe. This patch now adds
a define for the L2 cache IRQ number to the PPC headers and the l2_enable
routine uses this define.

Signed-off-by: Stefan Roese <sr@denx.de>