ppc: Make 440SPe PCIe code more generic to use on different 4xx PPCs
This patch is the first patch of a series to make the 440SPe PCIe code
usable on different 4xx PPC platforms. In preperation for the new 405EX
which is also equipped with PCIe interfaces.

(1) This patch renames the files from ppc440spe_pcie to ppc4xx_pcie

(2) This patch renames the functions from ppc440spe_ to ppc4xx_ with a
    little additional coding style cleanup

(3) This patch adds PCIe support for the 405EX to the common 4xx PCIe
    driver ppc4xx_pcie. After the previously done cleanups, this patch now
    adds most of the needed 405EX specific code. Also the Kilauea platform
    file is updated with the PCIe support.

(4) Another part of this patch is the introduction of PCIe IO space
    support. I'm not sure though how I could test this, since I don't have
    a PCIe board where the IO-space is used right now. The original code
    for this IO space init was done by AMCC for the Pine 405EX bringup
    board.

(5) This patch adds endpoint support for the AMCC Kilauea eval board. It
    can be tested by connecting a reworked PCIe cable (only 1x lane
    singles connected) to another root-complex.
    In this test setup, a 64MB inbound window is configured at BAR0 which
    maps to 0 on the PLB side. So accessing this BAR0 from the root-complex
    will access the first 64MB of the SDRAM on the PPC side.

(6) This patch also adds dynamic root-complex/endpoint configuration
    support. This is done by detecting the current configuration of each
    PCIe port, and reconfiguring this port accordingly. This way, the user
    can configure the port type in U-Boot via the "pcie_mode" env variable,
    and Linux will re-use this configuration again.

(7) On Yucca & Katmai, the inbound memory map pointed to 0x4.0000.0000,
    which is the internal SRAM. Since I now ported and tested this endpoint
    mode on Kilauea successfully to map to 0 (SDRAM), I also changed this
    for Katmai.
    Yucca will stay at internal SRAM for now. Not sure if somebody relies
    on this setup.

(8) Only reset the PHY when no link is currently established.
    This is for the Atheros PCIe board which has problems to establish
    the link (again) after this PHY reset. All other currently tested
    PCIe boards don't show this problem.

Signed-off-by: Stefan Roese <sr@denx.de>