ppc: Fix PCIe IRQ routing on Kilauea/Makalu
When a PCIe card is plugged into slot 1 and slot 0 has
not card, then the current IRQ routing is incorrect. The
card in PCIe slot 1 gets the IRQ from slot 0 assigned.

This patch fixes the problem and now assigns correct
IRQ number for both PCIe slots regardless which and where
the boards are plugged.

Signed-off-by: Stefan Roese <sr@denx.de>