Add touch_nmi_watchdog() calls to calibrate_delay()
This is needed for the Liebherr LWMON5 platform (not yet included
in kernel.org), which is equipped with a short timeout watchdog
(approx. 100ms). And this watchdog is already active upon Linux
bootup and can't be disabled. The time spent in calibrate_delay()
is too long without these touch_nmi_watchdog() calls resulting
in a board reset.

Note: The LWMON5 board is currently an arch/ppc platform available
in the DENX linux repositore (upstream merge of arch/ppc is not
possible anymore because it's scheduled for removal). A port to
arch/powerpc in the future is very likely to happen.

Signed-off-by: Stefan Roese <sr@denx.de>