md-raid-stripe-queues: fix an occasional deadlock in raid5
This is the revision of the "md: fix an occasional deadlock in raid5"
patch from Neil Brown for the Linux RAID driver which uses stripe
queues:

raid5/6's 'make_request' function calls generic_make_request on
underlying devices and if we run out of stripe heads, it could end up
waiting for one of those requests to complete. This is bad as
recursive calls to generic_make_request go on a queue and are not
even attempted until make_request completes.

So: don't make any generic_make_request calls in raid5/6 make_request
until all waiting has been done. We do this by simply setting
STRIPE_HANDLE instead of calling handle_stripe().

This change by itself causes a performance hit. So add a change so
that raid5_activate_delayed is only called at unplug time, never in
raid5. This seems to bring back the performance numbers. Calling it
in raid5d was sometimes too soon...

Signed-off-by: Yuri Tikhonov <yur@emcraft.com>