[Motion-PRO] Synchronize the LED dirver with the upstream state
This patch adds to the Motion-PRO driver changes resulting from its review
on the linuxppc-dev mailing list. In particular, it fixes a build issue
with the driver due to changes in OF access methods that went into the
mainline and then got merged with linux-2.6-denx. It's mostly based on the
(not-merged) usptream patch "[POWERPC] Add LED driver for Promess
Motion-PRO board.". See:
http://patchwork.ozlabs.org/linuxppc/patch?q=motion-pro&id=16617

Signed-off-by: Marian Balakowicz <m8@semihalf.com>
Signed-off-by: Bartlomiej Sieka <tur@semihalf.com>