USB: Fix problem with messages bigger than 64 bytes in DWC_OTG device driver
This patch fixes a problem in the device part of the DWC_OTG driver. Now
messages which exceed the frame size of 64 bytes will be sent and don't
lead to a timeout on the USB host side. Tested successfully on Canyonlands
(460EX) with USB ethernet and USB file storage gadget.

Signed-off-by: Stefan Roese <sr@denx.de>