PPC440SP(E) ADMA: fix processing the PQXOR operations.
Now, having ACK moved to flags, we should be carefull with this field of
the descriptor. So, when prepare we should set-up the flags obtained from
user only for the head descriptor of the chain. As for the rest - these
should be under internal control of the ADMA driver (for example, the
ppc440spe_adma_alloc_slots() function pre-ack all but last descriptors
using this fields, and so on).

Signed-off-by: Yuri Tikhonov <yur@emcraft.com>