x86: Fix root state corruptions
Only fair: I touched this code a few times recently while
re-enabling KGDB support, but I didn't noticed this lethal bug. Now I
had to hunt it for half a day.

What was wrong should be obvious. Its effect was a spuriously unstalled
root domain once in a while after we handled a fault over a non-root
domain. Fortunately this is a corner case - unless you use the FPU in
userspace (initial usage fault...).

Patch is (still) against 2.6.26, but it is required for all recent
i386/x86_64/x86 ipipe versions, from 2.6.23 to 2.6.27.

Signed-off-by: Jan Kiszka <jan.kiszka@web.de>