XSysAce: add new compatible, and 256K-related work-around
Add one more compatible string to the table for
of_platform binding, so that the platforms, which
have the SysACE chip on board (e.g. Katmai), could
describe it in their device trees correctly.

Also this patch updates the katmai.dts file with the
SysAce node, and makes the latter workable for the
case of 256KB PAGE_SIZE.

Signed-off-by: Yuri Tikhonov <yur@emcraft.com>
Signed-off-by: Ilya Yanok <yanok@emcraft.com>