raid: limit stripe cache size to avoid OOM problems
On platforms with PAGE_SIZE set to rather big values (e.g. 256KB) the
size of memory necessary for the stripe cache may be rather significant,
so we can't rely on the hard-coded value of initial max_nr_stripes
(this may cause invoking the OOM-killer for terminating RAID creation
process).

This patch checks the size of available memory, and assignes an appropriate,
safe, value to initial max_nr_stripes: either the hard-coded NR_STRIPES value
if it's safe in the sense that we'll have some free memory available using
stripe cache with such size, or the calculated, smaller, value.

Signed-off-by: Yuri Tikhonov <yur@emcraft.com>
Signed-off-by: Ilya Yanok <yanok@emcraft.com>