Prevent from CPU migration when accessing per-CPU domain variables
Unless specifically protected, code accessing ipipe_percpu variables
may be subject to CPU migration when running over the root domain in
preemptible sections, leading to all sorts of potential issues, as a
result of referring to the wrong per-CPU variable instance.

The proper access protection has been added in various locations
lacking it.  NOTE: low priority domains appearing after the root one
in the pipeline do refer to ipipe_percpu variables as well, therefore
we have to stop preemption by disabling hw interrupts, and we must do
this even when CONFIG_PREEMPT is disabled.

Kernel warnings are now issued when unsafe accesses are detected. This
detector is automatically compiled in when CONFIG_IPIPE_DEBUG and
CONFIG_SMP are enabled.