FCSE: Conversions between VA and MVA
Add to asm/fcse.h the functions needed to allow conversion between
virtual addresses (a.k.a VA) and modified virtual addresses (a.k.a
MVA), that is virtual addresses translated by the FCSE pid.

Though the whole kernel handles VA, and does not see generally MVA,
there are a few exceptions:
- the functions handling page table walking must use MVA, modifying
pgd_offset is enough;
- the TLB and cache operations may be passed user-space addresses must use
the MVA;
- the function do_DataAbort is passed an MVA which has to be converted
into a VA.

The cache handling is further modified. A bitfield keeps track of which
FCSE pid is dirty (that is, may have data present in cache), the tests of
mm_cpumask for cache operations are replaced with a test of this dirty bit,
since cache will not be flushed during context switches, and so the cache
may contain data from a process even though this process' mm is not the
current one. This bitfield is cleared upon full cache flushes in order to
avoid useless flushes.

Signed-off-by: Richard Cochran <richard.cochran@omicron.at>
Signed-off-by: Gilles Chanteperdrix <gilles.chanteperdrix@xenomai.org>