"Best-effort" FCSE: choose whether to flush cache at run-time.
The implementation of the FCSE "best-effort" mode requires to decide,
when switching contexts, whether to flush the cache. So, we add a
parameter to the mm switching function telling whether we want to
flush the cache, and the switch_mm macro uses the fcse_needs_flush
to get the value of this parameter.

Signed-off-by: Gilles Chanteperdrix <gilles.chanteperdrix@xenomai.org>