"Best-effort" FCSE: choose whether to flush cache at run-time
The implementation of the FCSE "best-effort" mode requires to decide,
when switching contexts, whether to flush the cache. So, we add a
boolean parameter to the mm switching function telling whether we want to
flush the cache. Modify the core-specific context switching functions
to only flush cache if this parameter is true.

Signed-off-by: Gilles Chanteperdrix <gilles.chanteperdrix@xenomai.org>