"Best-effort" FCSE: Handle shared mappings
Account for writable shared mappings, so as to be able to flush cache
when switching out a process with such a mapping.

This accounting is done by adding the L_PTE_SHARED flag to page
belonging to such mappings in make_coherent(), then, set_pte_at() is
intercepted to account for both addition and removal of pages with
this flag set.

Note that using the L_PTE_SHARED flag makes the FCSE patch
incompatible with ARMv6+ and Xscale3, but the FCSE feature is
deprecated in ARMv6 anyway.

Signed-off-by: Gilles Chanteperdrix <gilles.chanteperdrix@xenomai.org>