"Best-effort" FCSE: Handle mappings above 32 MB
In case an address larger than 32 MB is returned by
arch_get_unmapped_area(), relocate the calling process to the null
pid. When such a process forks, also give the child process a null
pid. Account for pids used by pages above 32 MB. The last process
with such pages is kept in cache as long as no other process uses
one of the PIDS it uses. When this happens, flush the cache.

We use set_pte_at() to do the pages accounting.

Signed-off-by: Gilles Chanteperdrix <gilles.chanteperdrix@xenomai.org>