x86: Fix root domain state restoring on exception return
If we enter __ipipe_handle_exception over a non-root domain and leave it
due to migration in the event handler over root, we must not restore the
root domain state so far saved on entry. This caused subtle pipeline
state corruptions. Instead, only save and restore them if we were
entering over root.

However, the x86-32 regs.flags fixup is required nevertheless to take
care of mismatches between the root domain state and the hardware flags
on entry. That may happen if we fault in the iret path. But also in this
case we must not restore an invalid root domain state. So if we entered
over non-root, pick up the input for __fixup_if from the root domain
after running the ipipe handler.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>