ipipe: wired interrupts cannot be locked
The interrupt locking flag is used to prevent a hw-originated
interrupt currently pending in the current domain's pipeline log from
being played, after it has been masked at PIC level. i.e.

stall(domain)
	<IRQ> -> deferred, pending in log
mask(irq)
	ipipe_irq_lock()
unstall(domain)

In that scenario, we don't want the interrupt to be played, since it
was masked after the domain was stalled. A correct PIC emulation
should keep the interrupt deferred, until it is unlocked by the
unmasking code. This is the purpose of the LOCK bit.

With a wired IRQ, which is always delivered to the head domain, such
precaution does not apply, since stalling the head domain actually
means masking interrupts.