ipipe: introduce return-to-userland intercept event
This patch introduces the generic bits for providing a mean to
dispatch IPIPE_EVENT_RETURN to domains, right before the kernel
returns to user-space, after an interrupt or a syscall.

This new event can be caught via ipipe_catch_event(). To reduce
overhead, it will be dispatched once, the next time the underlying
context exits to userland, after ipipe_return_notify() was called. The
event handler is passed a reference to the actual register frame which
will be unwound to resume user-space work.

A typical use of this feature consists of scheduling a callout on
behalf of an interrupt handler via ipipe_return_notify(), to do some
late fixups from the notification handler, right before the userland
context resumes.

IPIPE_EVENT_RETURN is dispatched with hw IRQs off.