ipipe: add debug checks to IRQ bitmap updates
Fast accessors to the pending IRQ bitmap (e.g. __ipipe_set_irq_pending
and callers) require local interrupts to be hard disabled. Failing to
do so may introduce nasty races - e.g. when interrupts are piling up
-, causing the IRQ bitmap to be wrongly updated, which usually ends up
in losing IRQ notifications.

This patch introduces assertions in those routines, to pull the break
when such a non-protected access is detected. This debugging code is
compiled in only when CONFIG_IPIPE_DEBUG_INTERNAL is enabled.