ipipe: remove "virtual irq only" pipeline sync option
Restricting the pipeline sync operation to virtual interrupts only
introduces the risk of deferring a pending IRQ until the log is synced
anew, which might not even happen at all (depending on the nature of
the interrupt "stuck" into the the log).

Typically, we could have:

-- __ipipe_sync_stage(virtual-only=1)

   for (;;) {
      irq = __ipipe_next_irq(virtual-only=1);
      if (no-more-virt-irq)
      	  break;
      <run IRQ handler>
      	        ...
         	__ipipe_schedule_irq(real-irq-number)
		...
    }

   /* exiting the syncer with real IRQ _not_  played */

Receiving a real interrupt over the root domain while hw interrupts
are on when running handlers would beget the very same issue. That
interrupt would be marked as pending into the log, but not played by
the syncer upon return from IRQ preemption. Alas, we do depend very
much on the assumption that the syncer _does_ always play all
interrupts which may be pending on entry to the call, or added to the
log while taking incoming IRQs over root domain handlers (note: taking
IRQs to mark them as pending, not dispatching them yet).

Current implementations only use the virtual-only sync mode when
calling the syncer from __ipipe_syscall_root(), which is wrong, and
will be fixed in the relevant arch-dependent sections as well.