x86/ipipe: fix random freeze at early boot (no console output)
Recent bootloader releases (>= grub 1.92 and grub2 for sure) may
introduce opportunities for device interrupts (typically disk
controller IRQs) to remain pending before control is passed to the
loaded kernel. In such a situation, telling the CPU to mask all
interrupts in the only way to protect the kernel from taking an
interrupt, particularly during the early boot phase when the interrupt
vectors are not set up yet.

Unfortunately, the interrupt pipeline tends to re-enable interrupts as
soon as possible, which includes forcing sti on the return path from
exceptions, assuming that our virtual "stall" bit would protect the
kernel from unwanted preemption just as well.

Alas, things break badly when a fault is taken during the early boot
phase mentioned earlier, causing the CPU to accept interrupts on the
return path of the exception. This is typically what happens when the
WP bit is tested in mem_init() for x86_32 platforms, which normally
triggers a page fault way before init_IRQ() had a chance to set up the
interrupt vectors properly. So, we end up with some pending interrupt
being taken inadvertently, causing execution to jump to ... nowhere.

To sum up, this bug requires three conditions to be met:

- running linux 2.6.31 or later, which introduces the initialization
  of the mm sub-system prior to initializing the interrupt vectors
  (i.e. mm_init() before init_IRQ()), which opens the race window.
- running a recent bootloader which is prone to leave interrupts
  pending before jumping to the kernel entry point (it seems lilo and
  old grub releases are not).
- running in 32bit mode. So far, the trigger found for this bug was
  the WP bit check, which is only applicable to 32bit platforms.

NOTE: this fix supplements commit a250e984a, which did not fully plug
the hole.