ipipe: x86-64: Update root pipeline state on TRACE_IRQ_ON/OFF
As we do not virtualize the IRQ flag in 64-bit entry code, the root
pipeline state gets out of sync. But lockdep applies some internal
consistency checks and fails for this reason. Fix this by conditionally
the state in the TRACE_IRQS_ON/OFF macros if tracing is enabled. This
requires to move some TRACE_IRQS_OFF spots as the original pipepline
state is otherwise overwritten before __ipipe_handle/divert_exception
can capture it.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>