ipipe: Add root preemption notifier
This is a per-cpu notifier mechanism, useful for informing ipipe-aware
code in the root domain about an ongoing preemption by a higher
prioritized domain. The first and likely only user will be KVM that
has to invoke some host state restorings on root preemption. Therefore,
this mechanism is single-user only for now.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>