ipipe: x86-32: Drop preemption prevention for ret_from_intr
This code prevented any legal preemption of kernel contexts under
CONFIG_PREEMPT. According to Philippe, "this hack was introduced in
2.6.9; disabling preemption on the IRQ return path was working around a
SMP issue causing a crash. I would not be surprised if such code was
only papering over a migration issue in fact, which probably disappeared
later in the pipeline implementation. x86_64 never required this, and it
is likely time to re-assess the need for this hack, and fix the issue
properly if ever it still makes sense."

No negative side effects were found when removing it, so let's do it
officially.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>