fcse: fix destroy_context
It seems that when init_new_context fails, destroy_context is called
for the same mm. This caused fcse_pid_free to be called, which declared
a potentially used pid as free, causing unexpected segmentation faults. Fix
this by using an unused pid value (127) to mark a failed new context, and
avoid calling fcse_pid_free when the mm passed to destroy_context has this
pid.