blackfin/ipipe: run the irq tail code unconditionally in syscalls
We might have a pending rescheduling when a high-priority domain is
issuing a syscall, in which case we do want to reschedule first, just
like we do for regular Linux syscalls.

This patch removes the dependency on the syscall value to call the irq
tail hook, so that it is called unconditionally.