ipipe: detect damage of hw interrupt state via spin_unlock_irqrestore
This patches catches a nasty issue where spin_unlock_irqrestore() on a
regular kernel spinlock is about to re-enable hw interrupts in a
section entered with hw irqs off. This is clearly the sign of a
massive breakage coming. Usual suspect is a regular spinlock which was
overlooked, used within a section which must run with hw irqs
disabled.