x86/ipipe: detect conflict with AMD's C1E power state
C1E power state is severely broken on AMD CPUs exhibiting erratum
halted state for idling.

The current kernel workaround is to switch the timers to broadcast
mode over the PIT directly from the idle loop, which causes the LAPIC
timers to be shut down and therefore unusable as tick devices.

This in turn causes __ipipe_request_tickdev() to fail when the LAPIC
is requested as the preferred tick device from a high priority domain,
leaving most callers in the cold.

In conclusion:

- one may pass idle=mwait on the boot cmdline to force the use of the
  P4-originated mwait/monitor instructions in the idle loop, thus
  preventing C1E state to be ever entered. Higher latencies were
  observed on some systems with those instructions though.

- preferably, C1E should be disabled at BIOS level, so that it does
  not interfere with real-time operations at all, and do not cause the
  LAPIC timers to be shut down by the kernel as a workaround.

This patch gives some hints to the user for fixing the issue when it
is detected.