"Best effort" FCSE: Reserve address-space for stack
In order to avoid a stack overflow which was occuring for processes
mapping to many libraries below the stack before using it, whereas in
best effort mode, mapping above 32MB is still possible, reserve, in
the address-space, the room for the stack size given by the rlimit.

Also decrease the default stack size rlimit to 1MB when FCSE is enabled,
as 8MB is a bit large if we want as much processes as possible using
a 32MB address-space.

Signed-off-by: Gilles Chanteperdrix <gilles.chanteperdrix@xenomai.org>