ipipe: expose two fixed domains for pipelining
Maintaining a variable list of domains prevents specific optimizations
when all pipeline stages are known in advance, and also induces some
overhead on the cache subsystem on low end hardware.

The use case we want to focus on involves a single high priority
domain, on top of Linux. Therefore we restrict the number of domains
to a fixed amount of 2, one for the regular kernel, and a second one
for a single client domain.