ipipe: introduce ipipe_root_only()
With only two possible domains in the pipeline, checking for call
boundaries via ipipe_check_context() boils down to make sure that a
Linux service is always called from the root domain.

For this reason, we move ipipe_check_context() to the legacy API, and
introduce ipipe_root_only(), which is a strict equivalent of the
former ipipe_check_context(ipipe_root_domain).