ipipe: remove stickiness bit, IRQ flags parameter from ipipe_request_irq()
We could only turn the sticky bit on with this parameter, but
stickiness is really a very peculiar property dedicated to handling
our critical IPI. Exposing it is useless at best, and dangerous at
worst.

Since we have no other flags and likely won't have any in the future,
we get rid of this parameter.