ipipe: remove ipipe_suspend_domain() from the client API
The head domain can only suspend by returning from any interrupt or
event handler it is executing, which means that only the root domain
may suspend explicitly from the kernel idle loop.

So there is no point in exporting an API call to only suspend the root
domain, this should be done via an internal service
(i.e. __ipipe_idle()), which boils down to unstalling the root stage
to flush the pending IRQs.