ipipe: introduce ipipe_post_work_root()
This is a mechanism to post a deferred work to the root domain from
the heading one. The work is defined as an argument structure passed
to a handler, both user-defined.

A struct ipipe_work_header shall head the argument structure passed to
ipipe_post_work_root(), mentioning the overall size of the data, and
the handler to execute within the root domain. The whole argument
structure is copied internally, and kept inside a private buffer until
the handler runs, so fairly small data should be sent this way not to
overrun the limited buffer space (2k per CPU).

The handler is guaranteed to run on the same CPU than the one
ipipe_post_work_root() was invoked from. It is legit to call this
service from the root domain as well: in such a case, the handler will
run immediately before ipipe_post_work_root() returns to the caller.

Typically:

void work_func(struct ipipe_work_header *work)
{
	struct foo *p = container_of(foo, struct foo, work);
	...
}

struct foo {
       struct ipipe_work_header header; /* MUST be first in data */
       int val;
       void *ptr;
} bar = {
	.header = {
		.size = sizeof(foo),
		.handler = work_func,
	},
	.val = some_integer;
	.ptr = some_pointer;
};

ipipe_post_work_root(&bar); /* Schedules the deferred call. */