ipipe: fix built in migration mechanism
The runqueue lock holder is not always the task reentering the root
context. This means that we must keep tracking the last outgoing task
from the root domain, to reinstate it as the Linux context owner when
reentering the root domain.

At this chance, we introduce a global per-CPU structure holding
CPU-local information for the pipeline core, which is not related to
domains.