ipipe: fold eoi+mask in fasteoi handler
As interrupt handlers may be deferred until the root stage is
unstalled which happens after the CPU accepts IRQs anew, we have to
mask the IRQ source for ackowledged fasteoi IRQs until their handler
have eventually run. This is required to prevent IRQ storms.

We now fold the EOI and masking operations in the pipeline's fasteoi
acknowledge handler, which removes the requirement for chip-specific
masking directly from the various irq_eoi() routines.